Extracted renderer structs in separate unit.
