+++ VulkanPrototype/src/Renderer/RendererUtils.cpp
@@ -1,0 +1,42 @@
+#include "RendererUtils.h"
+
+namespace VulkanPrototype::Renderer
+{
+    /*
+     * Member Functions
+     */
+
+    std::array<VkVertexInputAttributeDescription, 3> Vertex::getAttributeDescriptions()
+    {
+        std::array<VkVertexInputAttributeDescription, 3> attributeDescriptions{};
+
+        attributeDescriptions[0].binding = 0;
+        attributeDescriptions[0].location = 0;
+        attributeDescriptions[0].format = VK_FORMAT_R32G32_SFLOAT;
+        attributeDescriptions[0].offset = offsetof(Vertex, pos);
+
+        attributeDescriptions[1].binding = 0;
+        attributeDescriptions[1].location = 1;
+        attributeDescriptions[1].format = VK_FORMAT_R32G32B32_SFLOAT;
+        attributeDescriptions[1].offset = offsetof(Vertex, color);
+
+        attributeDescriptions[2].binding = 0;
+        attributeDescriptions[2].location = 2;
+        attributeDescriptions[2].format = VK_FORMAT_R32G32_SFLOAT;
+        attributeDescriptions[2].offset = offsetof(Vertex, textureCoordinate);
+
+        return attributeDescriptions;
+    }
+
+    VkVertexInputBindingDescription Vertex::getBindingDescription()
+    {
+        VkVertexInputBindingDescription bindingDescription =
+        {
+            .binding = 0,
+            .stride = sizeof(Vertex),
+            .inputRate = VK_VERTEX_INPUT_RATE_VERTEX
+        };
+
+        return bindingDescription;
+    }
+}
