+++ VulkanPrototype/src/Renderer/RendererUtils.h
@@ -1,0 +1,63 @@
+#ifndef RENDERERUTILS_H
+#define RENDERERUTILS_H
+
+#include <array>
+#include <fstream>
+#include <optional>
+#include <vector>
+
+#define GLM_FORCE_RADIANS
+#define GLM_FORCE_DEFAULT_ALIGNED_GENTYPES
+#include <glm/glm.hpp>
+#include <vulkan/vulkan.h>
+
+namespace VulkanPrototype::Renderer
+{
+    struct QueueFamily
+    {
+        std::optional<uint32_t> index;
+        uint32_t queueCount;
+    };
+
+    struct SurfaceDetails
+    {
+        VkSurfaceCapabilitiesKHR capabilities = {};
+        std::vector<VkSurfaceFormatKHR> formats;
+        std::vector<VkPresentModeKHR> presentModes;
+    };
+
+    struct UniformBufferObject
+    {
+        alignas(16) glm::mat4 model;
+        alignas(16) glm::mat4 view;
+        alignas(16) glm::mat4 proj;
+    };
+
+    struct Vertex
+    {
+        glm::vec2 pos;
+        glm::vec3 color;
+        glm::vec2 textureCoordinate;
+
+        static std::array<VkVertexInputAttributeDescription, 3> getAttributeDescriptions();
+        static VkVertexInputBindingDescription getBindingDescription();
+    };
+
+    struct UBOValues
+    {
+        //Model
+        float angle;
+        glm::vec3 axis;
+
+        //View
+        glm::vec3 eye;
+        glm::vec3 center;
+
+        //Proj
+        float fovy;
+        float near;
+        float far;
+    };
+}
+
+#endif // RENDERERUTILS_H
